refactor: Move `JsonRelationship` conversion to dedicated traits and modules.
